refactor: 이메일 조회 메서드 반환 타입을 Optional로 변경
- findByEmail 메서드의 반환 타입을 Optional<User>로 수정하여 null 처리 개선
