+++ src/main/java/com/market/saessag/domain/user/dto/SignUpRequest.java
@@ -1,5 +1,5 @@
 // 회원가입
-package com.market.saessag.domain.user.dto.request;
+package com.market.saessag.domain.user.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
